+++ lib/hobbit/environment.rb
@@ -1,5 +1,9 @@
 module Hobbit
   module Environment
+    def self.included(othermod)
+      othermod.extend self
+    end
+
     def environment
       ENV['RACK_ENV'].to_sym
     end